Add files via upload
Timestamp for reload config-message for log purposes.
Removed the print(msg) in the main loop to minimize spam.

Note, still no actual logging from the script.
Another observation is to do some better filtering of the main loop. All messages related to clock/transport seem redundant for now. This may be relevant to some use cases though, so maybe adding a config section to enable or disable this.